# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SmartCart AI - API Specification
  description: |
    Complete API documentation for SmartCart AI platform - an AI-powered supermarket ordering system.
    
    ## Architecture
    - Microservices-based architecture
    - RESTful APIs with JSON payloads
    - JWT-based authentication
    - Rate limiting enabled
    
    ## Base URLs
    - Production: `https://api.smartcart.ai`
    - Staging: `https://api-staging.smartcart.ai`
    - Development: `http://localhost:8080`
    
  version: 1.0.0
  contact:
    name: SmartCart AI Team
    email: api@smartcart.ai
  license:
    name: Proprietary
servers:
  - url: https://api.smartcart.ai/v1
    description: Production server
  - url: https://api-staging.smartcart.ai/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /nlp/parse:
    post:
      tags:
        - NLP
      summary: Parse natural language message
      description: Extract intent and entities from user message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - user_id
              properties:
                message:
                  type: string
                  example: "I need 2 kg tomatoes and 1 liter milk"
                user_id:
                  type: string
                  format: uuid
                context:
                  type: object
                  description: Conversation context
      responses:
        '200':
          description: Message parsed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      intent:
                        type: string
                        enum: [add_to_cart, view_cart, checkout, track_order, search_product, help]
                        example: "add_to_cart"
                      entities:
                        type: array
                        items:
                          type: object
                          properties:
                            product:
                              type: string
                            quantity:
                              type: number
                            unit:
                              type: string
                      confidence:
                        type: number
                        example: 0.95
                      suggested_response:
                        type: string

  # ==================== ANALYTICS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify-otp
...